cpu_input_boost: Avoid unnecessarily queuing work
The input handler wastes over 15 function calls every 100ms of constant
input. Queuing work when work is already queued locks local irqs and
fetches an atomic, regardless if it will decide to queue work or not. We
could potentially be wasting processing power here.

As a solution, we can have a boolean act as a basic lock to prevent the
input handler from trying to queue work if it is already queued, without
the overhead of irq locking and atomics. We only want it in the input
handler because this is the main source of arbitrary work queuing.

Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>